requirements.txt
django==5.0.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
django-environ==0.11.2
psycopg2-binary==2.9.9
redis==5.0.3
django-redis==5.4.0
celery==5.3.6
django-celery-results==2.5.1
django-cors-headers==4.3.1
drf-spectacular==0.27.1
bcrypt==4.1.2
cryptography==42.0.5
python-dotenv==1.0.1
uvicorn==0.29.0
gunicorn==21.2.0
requests==2.31.0
whitenoise==6.6.0
numpy==1.26.4
pydantic==2.6.4
fastapi==0.110.0
alembic==1.13.1
sqlalchemy==2.0.29
aiokafka==0.10.0
pydantic-settings==2.2.1
httpx==0.27.0